refactor(goxi): split the launcher mount-syscall proofs into their own binary
The widened derivation pushed launcher_child_filesystem_reachability.rs to
50243 bytes -- under the 51200 guard, but with under 1KB of headroom, so the
next table entry would have tripped it. Split rather than mark oversize.

The two halves are different kinds of evidence anyway: the reachability file
derives WHICH paths a brokered child depends on (across image ENV, rendered
worker env and the sandbox's spawn-time override) and classifies them against
the manifest; launcher_child_mount_syscalls.rs takes the mount set the renderer
produced and runs the real chdir(2) and File::create through a materialized
tree, with the workspace-mount-stripped ENOENT reproduction as its control.
44924 and 9628 bytes.
